requirements.txt
# Complaint Routing System - Dependencies
# Python 3.10+ required

# Data & Numerical Computing
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning & Models
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
joblib>=1.3.0

# NLP & Text Processing
sentence-transformers>=2.2.2
transformers>=4.30.0  # For sentence-transformers
huggingface-hub>=0.16.0

# Vector Search
faiss-cpu>=1.7.0  # Use faiss-gpu>=1.7.0 for GPU acceleration

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# Video Processing
opencv-python>=4.8.0

# API & HTTP
Flask>=2.3.0  # Optional, for REST API
fastapi>=0.100.0  # Optional, for async API
uvicorn>=0.23.0  # Optional, for FastAPI server

# Data Validation & Serialization
pydantic>=2.0.0
marshmallow>=3.20.0

# CLI
click>=8.1.0
typer>=0.9.0
rich>=13.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6  # For colored terminal output on Windows

# Optional: Data Exploration & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Development & Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0

# Jupyter for notebooks (optional)
jupyter>=1.0.0
ipython>=8.14.0
notebook>=7.0.0
